+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [v0.3.0-next.0](https://github.com/zalopay-oss/backstage-grpc-playground/tree/v0.3.0-next.0) (2022-08-29)
+
+[Full Changelog](https://github.com/zalopay-oss/backstage-grpc-playground/compare/v0.2.0-next.0...v0.3.0-next.0)
+
+**Implemented features:**
+
+- Support proto library [\#9](https://github.com/zalopay-oss/backstage-grpc-playground/issues/9)
+
+## [v0.2.0-next.0](https://github.com/zalopay-oss/backstage-grpc-playground/tree/v0.2.0-next.0) (2022-08-04)
+
+**Implemented enhancements:**
+
+- Support SSL [\#1](https://github.com/zalopay-oss/backstage-grpc-playground/issues/1)
+
+[Full Changelog](https://github.com/zalopay-oss/backstage-grpc-playground/compare/v0.1.3b...v0.2.0-next.0)
+
+## [v0.1.3b](https://github.com/zalopay-oss/backstage-grpc-playground/tree/v0.1.3b) (2022-07-19)
+
+[Full Changelog](https://github.com/zalopay-oss/backstage-grpc-playground/compare/v0.1.3...v0.1.3b)
+
 ## [v0.1.3](https://github.com/zalopay-oss/backstage-grpc-playground/tree/v0.1.3) (2022-07-15)
 
 [Full Changelog](https://github.com/zalopay-oss/backstage-grpc-playground/compare/v0.1.2...v0.1.3)